Map `while` nodes.
Since `while` can have their body precede their condition, I had to
make `mergeLocations` not automatically assume that the left node comes
before the right node. Also, this works around a bug in the
CoffeeScript parser that has the `column` value point to the start of
the `while` keyword instead of the start of the body for `a while a`
nodes.

https://github.com/jashkenas/coffeescript/issues/4114