fix: properly remove whitespace within herestrings
It turns out that CoffeeScipt's algorithm for removing whitespace from
herestring has a number of subtle cases and bugs that decaffeinate wasn't
properly replicating. This commit should hopefully fix all of those.

This commit also adds a number of tests that compare decaffeinate parsing
directly to CoffeeScript parsing, so hopefully that should provide more
confidence that the two are really equivalent.

Closes #393.