fix: properly calculate location data in template strings followed by `}` (#45)
The `buildLastQuasi` function, which would close out the last quasi in the case
where a template string ends in an interpolation, would start at the end index
and search backward until a `}` character. But the "end index" here is actually
one past the end, so if the template string was immediately followed by `}`, it
would call that the start and generate incorrect location data that would later
crash decaffeinate. Starting the search at one before the end index should fix
this and always be correct.

Closes https://github.com/decaffeinate/decaffeinate/issues/396 .