fix: improve detection of quasis in string interpolations
Fixes https://github.com/decaffeinate/decaffeinate/issues/227

Previously, the code would look at the raw source code directly to distinguish
between a real string and a quasi. But coffee-lex already does this work, so it
makes more sense to just rely on that. So now we just look for a STRING_CONTENT
node, which corresponds to a quasi, and say that we're a quasi if we find one.