fix: make string parsing more lenient about quotes between interpolations
Fixes https://github.com/decaffeinate/decaffeinate/issues/228
Fixes https://github.com/decaffeinate/decaffeinate/issues/515

When parsing literals, we already had a special case that if there's a problem,
we assume that it might be an interpolated string quasi and run fallback code
instead, but that was only running for one type of error. Now, we fall back in
both potential cases that could fail.

This isn't really the ideal solution; ideally we'd probably get rid of
parseLiteral altogether in favor of the coffee-lex output, which is generally
more robust, but this fix is nice in that it's simple and fixes some crashes.