fix: properly parse do statements with params defaulting to themselves (#76)
Fixes #539

The code to add commas exposed an issue in the parser where `do (a=a, b=b) -> c`
would parse incorrectly. The decaffeinate-parser code would see the key and
value as being the same, and assume that the value was implicitly added by the
CoffeeScript parser, so the params would show up as identifier nodes in the AST,
which later confused decaffeinate. Now, we do a more thorough check to see if
the default param was implicitly added by the CoffeeScript parser or explicitly
specified.