refactor: always use the String node type for strings (#85)
Progress toward these issues:
https://github.com/decaffeinate/decaffeinate/issues/557
https://github.com/decaffeinate/decaffeinate/issues/556
https://github.com/decaffeinate/decaffeinate/issues/554
https://github.com/decaffeinate/decaffeinate/issues/406

This change reworks the string representation in the decaffeinate-parser AST to
better fit the new way of thinking about strings:

* A simple string like `"Hello world"` is seen as a template string with 0
  interpolations rather than as something different from a template string.
* Herestrings and regular strings generally just differ in how the lexer treats
  them, rather than them being fundamentally different types of nodes. The
  coffee-lex output was intentionally designed so that the same (or nearly the
  same) code path can be used to handle whitespace trimming in either case.

To move toward this approach, this change makes it so all types of strings have
the node type `String`. In particular, there is no longer a field for
`TemplateLiteral.padding`, since the same information is kept in the token
stream now (and it seemed more straightforward for decaffeinate to read the
information directly from the tokens than indirectly through
decaffeinate-parser). A new `Quasi` node type is where the actual string
contents live (even though I don't expect decaffeinate will ever use quasi
contents).

One nuance here is that I still want to keep the underlying JS value for
literals, e.g. an actual int 3 for a literal 3. So I kept `parseLiteral`, but
changed it to no longer handle herestrings, and rather than operating on the
actual code, it now operates on the code provided in the CoffeeScript AST.

Still to do after this is to refactor the string interpolation code so that it
can work with heregex templates. Also, the quasi positions are currently taken
from CoffeeScript's positions, which isn't exactly what I want (ideally in
`"a#{b}c"`, the first quasi would start at the `a`, but for now, it starts at
the open-quote).

BREAKING CHANGE: The TemplateLiteral and Herestring node types have been removed
in favor of String. The String node type now always has an array of quasis and
an array of expressions. A new Quasi node type has been introduced for string
quasis.