feat: implement parsing for heregexes with interpolations
Progress toward these issues:
https://github.com/decaffeinate/decaffeinate/issues/557
https://github.com/decaffeinate/decaffeinate/issues/341

Just like with strings, we use decaffeinate-parser to unify the two styles of
heregexes (ones with interpolations and ones without) into a single Heregex node
type, and a normal heregexes simply have 0 interpolations. The CoffeeScript
parser treats heregexes with interpolations as a call to the `RegExp` function
with `+` operations for the components, so we can add a special case for that
type of node and reuse the code for dealing with string interpolations.

BREAKING CHANGE: Heregex expressions now always use the new Heregex node type
rather than RegExp. In all cases, a Heregex node has an array of quasis and
expressions.