fix: don't crash when inline JS ends in a close-quote (#89)
Fixes https://github.com/decaffeinate/decaffeinate/issues/576

Previously, we used string matching to figure out what type of literal the node
was, but this can't work by itself because inline JS just includes the JS
contents as the value, so `'a'` has the same CoffeeScript AST representation
whether or not it is surrounded in backticks. Instead, we can look at the
coffee-lex tokens and see if the literal corresponds to a `JS` token.

While I was at it, I also changed the heregex case to use tokens instead of
string matching.