fix: improve performance when adjusting node bounds (#94)
Progress toward https://github.com/decaffeinate/decaffeinate/issues/612

Previously, we did a linear search from the end of the token list until we found
a token in the right place. Now, we start the search from just after the end of
the node. We still include the bounds as part of the predicate, since
`indexOfTokenNearSourceIndex` just gives a good upper bound but not necessarily
an exact starting point.