fix: use coffee-lex tokens when generating quasi ranges (#101)
Fixes https://github.com/decaffeinate/decaffeinate/issues/641

There was still some code that was using simple string matching to find
interpolation bounds, so this change uses tokens instead. Relatedly, when we saw
to expressions in a row, we assumed that the quasi range is empty, but it
sometimes takes up space in the source code even if whitespace removal makes the
resulting string empty. So we now set the bounds to be the end of the last
interpolation end to the start of the interpolation start. We need to get these
bounds right because decaffeinate uses them to determine which tokens to
consider for padding removal.