refactor: extract some basic mappers from `parser.js` as TypeScript (#102)
The only real mappers that are active so far are for terminal nodes, i.e. nodes that have no children like booleans, nulls, strings without interpolations, etc. In order to handle nodes with children we need to cover all cases of terminal nodes, the biggest of which is `Literal` handled by `mapLiteral`.

BREAKING CHANGE: This makes a couple changes to the generated AST. Nodes generated by mapper functions are now instances of subclasses of `Node`, the base node type. They have acquired `start` and `end` properties which are the values of the `range` tuple, which is deprecated and will eventually be removed. Also, every node now has a `virtual` property and most are `false`. Finally, regular expression flags are now represented by a class which has the single-character boolean properties as well as more descriptively-named ones.