refactor: introduce `mapValue` for handling values
The main problem with doing this refactor in pieces is that they are mutually exclusive code paths. This means that once a call into `mapAny` or any of the concrete mappers happens within parser.js, all descendants of that node must be mappable with that system. Since most node types that contains children can contain essentially anything, this makes it tricky to do anything but all-or-nothing.

In this PR, I introduce not only a `mapValue` and the assiociated `Value` node, but a way to fall back to the old way of mapping. This way, if the mapper fails (likely because it cannot handle a particular node type), it will unwind back to the place we entered the mapper code path and continue as before. The downside is that we have to keep both code paths until we can switch everything over.