fix: convert string objects in the CoffeeScript AST into string primitives
Progress toward https://github.com/decaffeinate/decaffeinate/issues/666

When the CoffeeScript lexer sees an identifier that is also a reserved word, it
will make it into a string object so that it can tack a `.reserved` property on
it. However, since string objects have subtle differences from strings (most
notably, `===` can't be used to compare contents), this caused problems much
later on in decaffeinate. This commit converts to a string primitive if
necessary in a few places where it seems like this may come up, although the
MemberAccessOp case is the only one where I know for sure that it can.

The new test correctly tests that the `memberName` value is a string primitive;
the comparison fails if it's a string object.