fix: represent regular expressions as a `Regex` node
BREAKING CHANGE: Previously, regular expressions were represented as `Identifier` nodes, which hasn't been a problem for decaffeinate because all we wanted to do was pass them through anyway. Now, however, we want to do some processing on them and so need to identify them separately. Not to mention the fact that they are not identifiers!