refactor: try every node with the new mappers with fallbacks (#121)
This will allow us to make incremental progress on moving to the mapper functions without doing much to parser.js for a bit. Also, I renamed `Return#argument` to `Return#expression` to match the existing property name.

The downside of this is that it will likely increase parse times as we attempt to map a bunch of nodes we can't handle yet and unwind the stack. I haven't attempted to test this theory, but my gut says that it's not a big deal.