refactor: add several mappers for various node types
I did a bunch at once to coalesce the type errors that I had to fix in decaffeinate-coffeescript into a single version bump (11→12). The only change in this commit that should affect the output in a meaningful way is that some `Block` nodes which were not previously considered inline now are considered inline. This is due to a change from looking at source code for a preceding newline to using the tokens from coffee-lex, which is hopefully more accurate.