refactor: restructure `ProtoMemberAccessOp` and `MemberAcccessOp` (#134)
BREAKING CHANGE: Previously, `ProtoMemberAccessOp` represented both `a::b` and `a::`. Now, it represents just `a::`, meaning that `a::b` is now a `MemberAccessOp` that wraps a `ProtoMemberAccessOp`. `MemberAccessOp#memberName` property has also been replaced by the `member` property, which is an `Identifier` node. decaffeinate needs the position information for the member anyway, and this allows it to discard the code that looks for it.