test: check examples with and without mappers (#137)
The approach of using fallbacks when necessary has allowed us to do the transition to mappers/TypeScript in pieces. This is good, but it has the downside that bugs in the fallback code may be masked by a mapper works with the example code but not with real code which triggers the buggy fallback. This commit runs all examples twice, once with mappers+fallbacks, the other with only fallbacks. Doing so ensures that both code paths work as expected. To run with only mappers (and thus see what is left to port), run with `USE_FALLBACK=false`.