refactor: add mapper for `Class` nodes (#140)
I decided not to make any changes to the AST as part of this change, though we may want to consider it in the future. Specifically: do we really want to have another node type for the constructor? Having the constructor and bound methods already extracted for easy reference in decaffeinate is nice so it may be worth keeping those properties. Finally, `name` and `nameAssignee` are not the same thing in CoffeeScriptRedux, but they are here. We should either remove one or do the work of determining the class name and put that information in the AST. Currently decaffeinate is doing that itself. Note that we can't make `name` an identifier because of things like this: `class A::`. In that instance, the name of the class should be `prototype`, but there's no AST node that we can hook onto for that unless we make a virtual one. I added a skipped test for a case where the name isn't an identifier.