fix: check the context field when mapping object initializer members
The expressions `{a: b} = c` and `{a = b} = c` are nearly identical to the
CoffeeScript parser, except that the first one has an assignment with context
object and the second does not have a context. The new mapper code wasn't
reading this field and always making an ObjectInitialiserMember, which was
wrong, for now, we just fall back to the old correct code.