fix: properly parse assignments within `do` expressions
Progress toward https://github.com/decaffeinate/decaffeinate/issues/637

The `do` construct has special behavior when it takes a function with params,
and also has a special case for default params. This also extends to when the do
expression is an assignment and the RHS is a function with params. To handle
this, we just repeat the same logic on the RHS of an assignment if necessary.

Also disable some special case code where we were incorrectly adjusting the
location data of do expressions.