fix: use the right index when finding the token to the left of comparison operators (#152)
If there is no space between the left operand and the operator, we actually pass the start index of the operator (because it's the same as the end of the operand). Since we intentionally skip this token, because we assume it belongs to the operand, we never end up finding the operator. This fixes it by looking for the last token of the left operand one index before its end.

Fixes decaffeinate/decaffeinate#807.