fix: respect herecomments when computing function bounds
Fixes https://github.com/decaffeinate/decaffeinate/issues/687

Previously, the bounds of a function body were shrunk to only include semantic
tokens, which could cause a block comment to get moved outside the body and be
improperly indented after the normalize step. Now, we include the comments as
part of the body.

This is a small change that gets this case to work, but it might be that in the
future we'll want to avoid all AST location adjustment and/or put block comments
in the AST.