fix: overhaul template literal parsing to rely on token positions (#186)
The previous code mostly relied on AST nodes and manual string matching, which
was fragile and didn't seem to provide a clean way to allow empty string
interpolations. The new approach is to walk the template literal, keeping track
of the coffee-lex tokens and defining the quasi ranges in terms of those. This
means that the first and last quasi no longer bleed into the quotation marks,
and other small location data fixes.

Empty string interpolations still aren't 100% working, but this commit gets most
of the way there.