refactor: change template literal handling to operate on CoffeeScript nodes (#187)
Previously, we would convert to decaffeinate-parser nodes, which also meant that
we had to know up-front when something would be a quasi. Now, we instead find
the unmapped template literal expressions and then map them later. This should
make it easier to handle some remaining cases of empty string interpolations,
and also provide a better path for moving all of the template literal code to
typescript/mappers.