fix: change string interpolation handling to fully handle empty interpolations (#189)
We now have helper functions as a unified place to handle string and heregex
parsing (on the mapper side). This also means that all strings, not just
interpolated strings, have correct quasi bounds now. In the previous case where
we were getting a "bare" quasi, we can now just make it a normal string, using
the tokens as the source of truth.