fix: change top-level program creation to use mapBlock
This fixes an issue where top-level comments weren't being filtered out
correctly.

Also fix some issues exposed by this change when computing the "inline" status
of a block. The mapper side needs to set post-if consequents as inline, and we
should not treat the block as inline if it's at the very start of the program.
Also fix post-for and post-while bodies to be inline, which was only happening
by accident before.