test: update test data to be canonical
Rather than using JSON.stringify, which prints keys in iteration order (i.e.
insertion order), we now use the json-stable-stringify package, which is just
alphabetical. I also added an OVERWRITE_EXPECTED_OUTPUT env variable flag that
makes it easy to clobbber the expected output with the actual output, which I
also intend to use in an upcoming diff.

This should make it a lot easier to manually compare changes and to more safely
make large-scale changes.