feat: add getChildren and getChildFields methods to Node (#249)
This should make it easier to traverse ASTs when depending only on
decaffeinate-parser.

Also change test traversal to use this new method, so that our tests make sure
every node is covered. (`stripExtraInfo` also had the wrong type before, so this
makes it so the param is actually always `Node`.)

Traverse function and parent pointers coming soon.