feat: add a getChildNames method
This is needed for some use cases where we want to use the name of the child,
not just the value.

I wanted to use `Array<keyof this>` everywhere, but it turns out that doesn't
work super-well with inheritance, so I can only use it for derived classes and
use `Array<string>` for base classes.