fix: add node 8 to the travis build (#270)
We need node 8 for semantic-release to work, and it's a good idea to add it
anyway. Note that this works with travis_after_all since the first build is
always the build leader, but it's possible we can/should get rid of that, but
I'll do the easy thing for now.

Also fix an issue with the lockfile being non-canonical, presumably due to a
greenkeeper issue.