feat: switch to the CS2 AST format (#317)
There's some CS2-specific code, but we don't yet ever run the CS2 parser.
Instead, we run the CS1 parser, convert that AST to a CS2 AST, then convert that
to a decaffeinate-parser AST. This should make it possible to variably use the
CS1 or CS2 parser depending on configuration.

BREAKING CHANGE: Node 4 is no longer supported. The returned CoffeeScript AST is
now a CS2 AST, and there are slight differences in ranges and AST structure for
comment-only blocks.