fix: improve CS1 function ranges (#318)
CS2 doesn't provide as much location information when a comment is at the end of
a function block, but also is less strict about block comment positioning. In
CS1, we need to be careful to end the function in a place that doesn't leave a
stray indented block comment, so we track CS1 comments and remove them in an
intermediate state so that function bounds are more accurate.