fix: expand block node locations to include surrounding parens (#334)
This fixes cases like `(break) while true` in CS2, where the new AST was
treating `break` as its own block and there was no source index to use to expand
the whole loop to start at the open-paren. Now, we always expand blocks to
include their surrounding parens, which seems to make sense in general.