fix: shrink empty blocks to have a size of 0 when fixing locations
We now expand blocks to include their surrounding parens, but that made
expressions like `(->)` break. The block in the arrow function had the range of
the `>` since CoffeeScript doesn't support empty blocks, so it was viewing the
empty block as being surrounded by parens. To fix, we can make the block
actually empty, so the token to the left is `>`, which seems to fix things.